modularize db controller helpers closes #101
